+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Changes since v2.29
   - In the future, autosave may become the default mode. (#2767)
   - Applications don't yet automatically get reloaded, should another person be viewing the same application.
     This is a potential future feature. (see #2622, #2244)
+- Data Use Ontology (DUO) codes are now rendered in application pdf. Rendered pdf styles have also been tuned for more consistent look. (#2857)
 - Organization owners are now allowed to edit their own organizations owners. (#2828)
 
 ### Fixes